Improvements to the xdelta code in order to take advantage of the fact that we are always using a relatively small window for the adler32 hash, and that when we skip forward, we normally skip forward by alot so it's faster to just recompute the rolling checksum on the new data than actually move the rolling checksum forward. 1.03x improvement in cpu usage on the client, 1.12x improvement in cpu usage on the server.
adler32.hh: move a bunch of things into being constants, and migrate
to just doing a direct cast of the character rather than the widen()
stuff because the widen does a bunch of unnecessary masking.  Code
added to xdelta.cc to verify correctness of this transform.  Optimize
the insertion of a bunch of values for the case where the count is
small; just put an invariant in place to require that condition as
it's the only one used.  For the small update case, we can skip a lot
of the masking that we would otherwise have to do as nothing can
overflow.

xdelta.cc: mark the blocksz as constant (it is), remove the unused hi
variable from compute_delta_insns.  Specialize the xdelta advancement
code to handle the case where we advance by a lot by skipping over the
intermediate characters rather than pulling the rolling checksum
forward the entire way.  Add in some simple code to verify that the
widen transform we made is valid.

Statistic                  ref-2245   xdelta
pull-avg-resident-MiB       46.69 ?    46.78
pull-avg-size-MiB           52.21 ?    52.20
pull-max-resident-MiB       65.32 <    65.89
pull-max-size-MiB           72.01 ?    72.22
pull-system-time            20.44 >    19.67
pull-user-time             676.79 >   657.13
pull-wall-time             919.58 >   895.51
server-avg-resident-MiB     53.41 ?    53.32
server-avg-size-MiB         58.66 ?    58.57
server-max-resident-MiB     59.56 ?    59.43
server-max-size-MiB         65.02 ?    64.98
server-system-time          19.55 <    20.61
server-user-time           190.12 >   166.96
server-wall-time           922.73 >   898.66