add support for both scgi and http connections
the command 'mtn --db server.db gserve' will start the server in scgi mode with
the expectation that it is behind an scgi capable web-server. the command 'mtn
--db client.db gsync http://localhost/monotone "*"' will start the client and
connect to the webserver in front of the monotone scgi server.

the command 'mtn --db server.db gserve --http' will start the server in http
mode without the requirement for an intermediate scgi web-server. the command
'mtn --db client.db gsync http://localhost:8008 "*"' will start the client and
connect directly to the associated monotone server.

* cmd_netsync.cc (gsync): add note about also supporting stdio
* cmd_scgi.cc: add connection_type enum; 
  (scgi_error): rename to...
  (gserve_error): ... this
  (parse_scgi): rename to ...
  (parse_scgi_headers): ... this
  (parse_http_headers): new function for parsing raw http headers
  (process_scgi_transaction): rename to ...
  (process_transaction): ... this; used for both http and scgi
  (CMD(scgi)): rename to ...
  (CMD(gserve)): ... this; remove unused scgi_bind option; add bind_http option
* constants.hh: set default_http_port to 8008 so non-root users can run server
* http_client.cc (transact_json, parse_http_status_line, parse_http_header_line,
  parse_http_response): upgrade to http 1.1 in an attempt to fix connection
  closing problems
* json_msgs.cc: add encode_hexenc and decode_hexenc calls in various places to
  account for binary revision and file ids
* options_list.hh: add bind_http option; remove unused scgi_bind option