Colorized output of mtn status
For simplicity I have made status_added be used when adding files 
and directories. We may want to differentiate between them later 
like we do for diff_* and status_* even though they both have very 
similar "child" purposes, but for now my plan is to simplify them 
and be consistent with their usage. 

This has the benefit of the get_output_color hook being easier to
edit by all users.

For instance, nobody should have to care about the "reset" purpose
in the hook as this is an implementation detail, so this will be
taken out eventually.