database::get_revision - invariant for revision's existence
Currently, some of the callers to get_revision manually call
revision_exists and handle the non-existence manually - usually
with the E macro.

The rest, along with get_revision just assume it works. For
instance get_revision assumes it works and then continues to
gzip decode the data - so we make the data returned from the
db an invariant.

This passes all tests as defined by "make check".