Created new revert_in_two_parent_workspace func test
Moved the existing revert test from two_parent_workspace_commands_that_fail
into it, and also copied the setup/writefile/commit template to produce a
workspace with 2 parents.

All the tests currently pass, however the new test is not yet finished so
I've left some TODO's in it for the meantime. The bottom section of
commented out code should be correct as it checks the workspace contents
after a revert to "right" - it just depends on the unimplemented revert to
"right" code.

Also, the code which checks the contents of the files should really be
"left\nancestor", but I couldn't get it to work by using \n in the string.
As a work around I check for both left + ancestor, and check for NOT right.