Extra checks for revert_in_two_parent_workspace
We now check stdout to make sure only the files we expect have been
reverting, as well as check the contents afterwards.

In addition, we check which revision the workspace should now be based off
and if there no changes. The base revision should be the one we reverted
to, and there should be no changes. This currently fails on line 50, with
the base revision check. I do not know why yet, as it seems to think the
base revision is "right", instead of "left".

Performing the same sequence of steps manually in a different terminal
session, all works correctly and the base revision is "left".