Use '<= items.size()' for error condition in 'selector::create'
When executing monotone with the reported selector, this was the error
guard causing monotone to fail. The values were lparen_pos = 2,
items.size() = 2.

I think the original use of '< items.size()' may have been a typo, as
a few lines earlier '<= items.size()' is used for finding the position
of the nearest previous '('.

The corresponding 'if' will now evaluate to false, as lparen_pos ==
items.size(), and line 796 will be executed. The error check there for
lparen_pos == 2 also ties in with the crash report of lparen_pos being 2.

The recently added 'selectors_with_parentheses' test now succeeds, along
with other selector related tests.