Append extra newline to the output of "mtn au inventory"
Using "cmd.exe" on Windows XP SP3, "mtn au inventory" output (and some
other automate commands) is blank; linux systems are fine, and apparently
Windows 7 is as well.

Other automate commands run on Windows, such as "mtn au leaves" are fine. A
quick glance at the code shows that "leaves" outputs a separate '\n', which
may be causing the output to be flushed.

This is done here 1st as an experiment to see if it's the cause for missing
output on Windows, and if it is, then more revisions will follow for
remaining commands.