Correct failing tests for the dropped/modified conflict resolution
The logic is fine, but the syntax used is ambiguous to Lua. See section 2.5.8
of the Lua 5.1 reference manual which states

> As an exception to the free-format syntax of Lua, you cannot put a line
> break before the '(' in a function call. This restriction avoids some
> ambiguities in the language. If you write
>
>    a = f
>   (g).x(a)
>
>
> Lua would see that as a single statement, a = f(g).x(a). So, if you want
> two statements, you must add a semi-colon between them. If you actually
> want to call f, you must remove the line break before (g). 

To fix, make sure we put () representing a function call on the same line
as the function call itself; in this case the call to 'samelines'.

This was found originally on a SUSE Linux Enterprise 11 SP2 system with
Lua 5.1.4 and this patch fixes it on that system. For posteriry it's
reported that the original code also succeeded on other systems (although
the Lua version was omitted), such as

+ Windows XP SP3 MingW & Cygwin
+ Windows 7 MingW & Cygwin (32 bit)
+ Debian testing