Make 'expand_date' always return a date string in UTC
This was found as perfoming a commit and then `mtn log -r"l:5 minutes ago"`
would return no matches instead of the just committed revision; tested in
the UK which is currently 1 hour ahead of UTC.

The problem was we were returning a local time value which was then compared
against the UTC value stored in the database. To fix always return a UTC
time from 'expand_date'. I've also updated the logging line to say the date
has been expanded to UTC format for clarity.

However, we could do with a test for this. I tried adding the
`mtn log -r"l:5 minutes ago"` case to 'check_later_and_earlier_selectors'
but the test still passed even though it fails in a terminal; this is because
the tests always run in UTC mode.